+++ spec/spec_helper_acceptance.rb
@@ -1,46 +1,6 @@
-require 'beaker-puppet'
-require 'puppet'
-require 'beaker-rspec'
-require 'beaker/puppet_install_helper'
-require 'beaker/module_install_helper'
-require 'beaker/testmode_switcher'
-require 'beaker/testmode_switcher/dsl'
+# frozen_string_literal: true
 
-run_puppet_install_helper
-configure_type_defaults_on(hosts)
-install_module_on(hosts)
-install_module_dependencies_on(hosts)
+require 'puppet_litmus'
+require 'spec_helper_acceptance_local' if File.file?(File.join(File.dirname(__FILE__), 'spec_helper_acceptance_local.rb'))
 
-RSpec.configure do |c|
-  # Readable test descriptions
-  c.formatter = :documentation
-
-  # detect the situation where PUP-5016 is triggered and skip the idempotency tests in that case
-  # also note how fact('puppetversion') is not available because of PUP-4359
-  if fact('osfamily') == 'Debian' && fact('operatingsystemmajrelease') == '8' && shell('puppet --version').stdout =~ %r{^4\.2}
-    c.filter_run_excluding skip_pup_5016: true
-  end
-
-  # Configure all nodes in nodeset
-  c.before :suite do
-    hosts.each do |host|
-      # Required for binding tests.
-      if fact('osfamily') == 'RedHat'
-        if fact('operatingsystemmajrelease') =~ %r{7} || fact('operatingsystem') =~ %r{Fedora}
-          shell('yum install -y bzip2')
-        end
-      end
-      on host, puppet('module', 'install', 'stahnma/epel')
-    end
-  end
-end
-
-shared_examples 'a idempotent resource' do
-  it 'applies with no errors' do
-    execute_manifest(pp, catch_failures: true)
-  end
-
-  it 'applies a second time without changes', :skip_pup_5016 do
-    execute_manifest(pp, catch_changes: true)
-  end
-end
+PuppetLitmus.configure!